Move builder to util crate
