+++ util/src/builder/mod.rs
@@ -1,3 +1,6 @@
 //! Builders for large structs.
 
+mod callback_data;
 pub mod command;
+
+pub use callback_data::CallbackDataBuilder;
